[Refactor] Refactor CCE update node
