+++ openstack/cce/v3/nodes/testing/requests_test.go
@@ -235,7 +235,7 @@ func TestUpdateV3Node(t *testing.T) {
 		_, _ = fmt.Fprint(w, Output)
 	})
 	options := nodes.UpdateOpts{Metadata: nodes.UpdateMetadata{Name: "test-node"}}
-	actual, err := nodes.Update(fake.ServiceClient(), "cec124c2-58f1-11e8-ad73-0255ac101926", "cf4bc001-58f1-11e8-ad73-0255ac101926", options).Extract()
+	actual, err := nodes.Update(fake.ServiceClient(), "cec124c2-58f1-11e8-ad73-0255ac101926", "cf4bc001-58f1-11e8-ad73-0255ac101926", options)
 	th.AssertNoErr(t, err)
 	expected := Expected
 	th.AssertDeepEquals(t, expected, actual)
