+++ openstack/cce/v3/nodes/urls.go
@@ -2,11 +2,6 @@ package nodes
 
 import golangsdk "github.com/opentelekomcloud/gophertelekomcloud"
 
-const (
-	rootPath     = "clusters"
-	resourcePath = "nodes"
-)
-
 func rootURL(c *golangsdk.ServiceClient, clusterid string) string {
 	return c.ServiceURL("clusters", clusterid, "nodes")
 }
